refactor: Change variable and file naming, also the response messages for login and logout
